# api/openapi.yaml
openapi: 3.0.3
info:
  title: Zyad Cloud API
  version: 0.1.0
paths:
...
  /api/v1/admin/landing-pages/{id}/branding:
    get:
      summary: Get landing page effective branding
      operationId: getLandingPageBranding
      tags:
        - Landing Branding
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Page branding retrieved successfully
    patch:
      summary: Update landing page branding override
      operationId: updateLandingPageBranding
      tags:
        - Landing Branding
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Page branding override updated successfully
    delete:
      summary: Delete landing page branding override
      operationId: deleteLandingPageBranding
      tags:
        - Landing Branding
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Page branding override removed successfully
  /api/v1/admin/landing/branding:
    get:
      summary: Get default landing branding
      operationId: getDefaultLandingBranding
      tags:
        - Landing Branding
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Default branding retrieved successfully
    patch:
      summary: Update default landing branding
      operationId: updateDefaultLandingBranding
      tags:
        - Landing Branding
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Default branding updated successfully
  /api/v1/admin/landing/theme:
    get:
      summary: Get landing theme tokens
      operationId: getLandingTheme
      tags:
        - Landing Branding
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Theme retrieved successfully
    patch:
      summary: Update landing theme tokens
      operationId: updateLandingTheme
      tags:
        - Landing Branding
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Theme updated successfully
  /api/v1/admin/landing/domains/available:
    get:
      summary: List available landing domains
      operationId: listAvailableLandingDomains
      tags:
        - Landing Domains
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Available domains retrieved successfully
  /api/v1/admin/landing/domain-bindings:
    get:
      summary: List landing domain bindings
      operationId: listLandingDomainBindings
      tags:
        - Landing Domains
      security:
        - BearerAuth: []
      parameters:
        - name: landing_page_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Domain bindings retrieved successfully
    post:
      summary: Create landing domain binding
      operationId: createLandingDomainBinding
      tags:
        - Landing Domains
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Domain bound successfully
  /api/v1/admin/landing/domain-bindings/{id}:
    patch:
      summary: Update landing domain binding
      operationId: updateLandingDomainBinding
      tags:
        - Landing Domains
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Domain binding updated successfully
    delete:
      summary: Delete landing domain binding
      operationId: deleteLandingDomainBinding
      tags:
        - Landing Domains
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Domain binding removed successfully
  /api/v1/admin/landing/ctas:
    get:
      summary: List landing CTAs
      operationId: listLandingCtas
      tags:
        - Landing Reusable
      security:
        - BearerAuth: []
      responses:
        "200":
          description: CTAs retrieved successfully
    post:
      summary: Create landing CTA
      operationId: createLandingCta
      tags:
        - Landing Reusable
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: CTA created successfully
  /api/v1/admin/landing/ctas/{id}:
    patch:
      summary: Update landing CTA
      operationId: updateLandingCta
      tags:
        - Landing Reusable
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: CTA updated successfully
    delete:
      summary: Delete landing CTA
      operationId: deleteLandingCta
      tags:
        - Landing Reusable
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: CTA deleted successfully
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...